backend/pyproject.toml
[project]
name = "perception-ai-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.87.0",
    "bcrypt>=5.0.0",
    "google-genai>=1.0.0",
    "openai>=1.60.0",
    "celery[redis]>=5.4.0",
    "fastapi>=0.115.0",
    "motor>=3.6.0",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.5.0",
    "redis>=5.0.0",
    "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
test = [
    "httpx>=0.28.0",
    "mongomock-motor>=0.0.30",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
]
dev = [
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.setuptools]
packages = ["app"]
